8-patterns/requirements.txt
langchain
langchain-openai
langgraph
pydantic
python-dotenv
typing-extensions
